🚚 move: store 폴더 생성, 패킷프로토 최신화
